+++ package.json
@@ -8,6 +8,8 @@
     "@mui/material": "^5.14.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router": "^6.17.0",
+    "react-router-dom": "^6.17.0",
     "use-debounce": "^9.0.4"
   },
   "devDependencies": {
